add scope test and renamed test class
