+++ mybatis-spring-boot-samples/mybatis-spring-boot-sample-annotation/pom.xml
@@ -45,6 +45,7 @@
 		<dependency>
 			<groupId>org.mybatis.spring.boot</groupId>
 			<artifactId>mybatis-spring-boot-starter-test</artifactId>
+			<scope>test</scope>
 		</dependency>
 	</dependencies>
 	<build>
