+++ mybatis-spring-boot-samples/mybatis-spring-boot-sample-annotation/src/main/resources/import.sql
@@ -1,5 +1,5 @@
 --
---    Copyright 2015-2016 the original author or authors.
+--    Copyright 2015-2017 the original author or authors.
 --
 --    Licensed under the Apache License, Version 2.0 (the "License");
 --    you may not use this file except in compliance with the License.
@@ -13,6 +13,8 @@
 --    See the License for the specific language governing permissions and
 --    limitations under the License.
 --
+
+drop table if exists city;
 
 create table city (id int primary key auto_increment, name varchar, state varchar, country varchar);
