+++ mybatis-spring-boot-samples/mybatis-spring-boot-sample-annotation/src/test/java/sample/mybatis/SampleMapperTest.java
@@ -13,51 +13,40 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-
 package sample.mybatis;
 
-import org.apache.ibatis.session.SqlSession;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mybatis.spring.boot.test.autoconfigure.MybatisTest;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.context.annotation.Bean;
 import org.springframework.test.context.junit4.SpringRunner;
-import sample.mybatis.dao.CityDao;
 import sample.mybatis.domain.City;
+import sample.mybatis.mapper.CityMapper;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 /**
+ *
  * @author wonwoo
  * @since 1.2.1
  */
 @RunWith(SpringRunner.class)
 @MybatisTest
-public class SampleSqlSessionTests {
+public class SampleMapperTest {
 
   @Autowired
-  private SqlSession sqlSession;
+  private CityMapper cityMapper;
 
   @Test
-  public void sqlSessionIsNotNullTest() {
-    assertThat(sqlSession).isNotNull();
+  public void mapperIsNotNullTest() {
+    assertThat(cityMapper).isNotNull();
   }
 
   @Test
-  public void selectCityByIdTest() {
-    City city = sqlSession.selectOne("selectCityById", 1);
+  public void findByStateTest() {
+    City city = cityMapper.findByState("CA");
     assertThat(city.getName()).isEqualTo("San Francisco");
     assertThat(city.getState()).isEqualTo("CA");
     assertThat(city.getCountry()).isEqualTo("US");
-  }
-
-  @TestConfiguration
-  public static class Config {
-    @Bean
-    public CityDao cityDao(SqlSession sqlSession) {
-      return new CityDao(sqlSession);
-    }
   }
 }
